+++ presentation/src/main/java/com/anytypeio/anytype/presentation/search/ObjectSearchConstants.kt
@@ -129,15 +129,6 @@ object ObjectSearchConstants {
             relationKey = Relations.IS_HIDDEN,
             condition = DVFilterCondition.NOT_EQUAL,
             value = true
-        ),
-        DVFilter(
-            relationKey = Relations.TYPE,
-            condition = DVFilterCondition.NOT_IN,
-            value = listOf(
-                OBJECT_TYPE_URL,
-                RELATION_URL,
-                TEMPLATE_URL
-            )
         )
     )